rdtsc timing; openmp loop; cleanup
Added an RDTSC-based timer to avx_add, although this requires explictly
setting the reported CPU frequency.

The unrolled loop was also removed, since GCC and ICC seem to be
reliably unrolling it.

----

NOTE: RDTSC returns a cycle count, but more like a time measurement in
units of the unscaled frequency (e.g. 2601MHz on Raijin).

If `T` is the runtime, then RDTSC returns `f_0 / T` cycles, where `f0`
is the unmodified frequency.  The actual number of cycles is given by
`f` / `T`, where `f` is the actual frequency due to CPU scaling effects
such as Turbo Boost or thermal throttling.

One needs the to know the CPU frequency in order to convert RDTSC into a
runtime.