axpy benchmark
Adding the `axpy` benchmark to microbench.

axpy is still incomplete, it is simply a test of y[:] = a * x[:] + y[:]
flops.  (sometimes without `a` depending on my testing)

I've also attached an AVX version to help out gcc, but this also needs
testing.

This is all very messy "work in progress" stuff.