Generalised AVX timers
avx_add was duplicated into `avx.c`, and a new timer object was defined
in `gettime.c`.

This is still very primitive, but it's some first steps towards removing
the RDTSC preprocessors and to support both timers in a single build.

Future steps:  Move the start/end/runtime functions into the TimeContext
objects, which will simplify things quite a bit.