Prototype polymorphic timer added
A prototype timer has been added that can act as both a TSC or a
POSIX-based timer.  It is currently saved under `ptimer.c`.

It's not integrated into the avx code yet, nor is it well organised.  It
can probably be cleaned up with some typedefs, and a proper header needs
to be split off.  But it works, and it is probably a good idea here.

Shamelessly ripped from this Stack overflow question:

    http://stackoverflow.com/questions/415452/object-orientation-in-c