Complete Timer rewrite
Timer was completely re-written to avoid the complicated polymorphing
from the previous version.  This new one is much simpler.  Just create
the number with an enum setting the type, then call `start`, `stop` and
`runtime`.  Internal data is now saved in an `union`, so it can differ
for different timer types.  But all of that is obscured from the user.

I am happy with this timer now, but the namespace is all wonky and
conflicts with the `time.h` stuff, so the internal variable names may
need to be changed.  But AFAIK this is all working.  YAY.