avx_add and avx_mac integrated into avx.c
The AVX and and Multiply-add-concurrent tests are no longer standalone,
and have now been integrated into the avx.c file.  Most of the work was
done in the previous commits; this commit removes the original avx_add.c
and avx_mac.c files.

This is a step towards integrating all of the vector tests into a single
executable.