flop: streamline avx tests with `volatile`
We now use the `volatile` keyword to preserve `result` and only save it
as an internal argument, rather than passing it as a pointer into the
function.

Preserving and printing `result` was only done to force the compiler to
preserve the register arithmetic involved.  Declaring `result` as
`volatile` forces the compiler to preserve the value, regardless of
whether or not it's used, and streamlines the code to some degree.

We might be able to make further improvements but declaring the register
variables as volatile, and eliminate the binomial sum of `result`.