Dynamic TSC frequency calculation
I have added a preliminary method for calculating the TSC frequency
dynamically, rather that hard-coding it into the source.

This is a very crude method; we call nanosleep for 1 second and then set
the TSC frequency to the difference of the rdtsc instruction output.
It's not robust to more than about four digits accuracy, but it's good
enough for now.

The best methods read the Master Boot Record, but this requires root
privieleges.  Better methods will call rdtsc multiple times in
succession to get a consistent value (or at least this is what I've seen
in various kernel tools), but I'm not yet prepared to write something
like that.