Potential OpenMP support; tsc cpufreq change
A possible framework for doing OpenMP runs was added, some of the
original layout has been cleaned up and the parallelisation was moved to
the flop.c driver file.  Results are a bit confusing, so this may not be
working correctly yet.

`timer_set_tsc_freq` was changed to `timer_get_tsc_freq` and is now only
calculates the TSC frequency, which is used by the initializer to set
`cpufreq`.