avx_mac repeat loop removed
The explicit double-unroll of the avx_mac (concurrent add-multiply) was
removed, because it was confusing the Intel compiler, which was trying
to save all of the intermediate values and was running out of AVX
registers.  (GCC was making no effort to save them.)

Intel and GCC now both produce something close to peak performance in
the avx_mac test.