Maximum core count check
I have added a check to ensure that the number of threads does not
exceed the maximum core count.  This was needed because CPU_* functions
would not assign a thread to a nonexistent CPU, and the pthread barrier
would wait for the unassigned threads.

The current processor count check uses OpenMP; there are other ways to
do this, but this one seems to be the most portable at the moment.