Dynamic loop iteration; local flop calculation
Loop iteration is now dynamic rather than a huge static constant.  The
new method doubles the number of iterations until the total runtime
equals 0.5 seconds.

This introduces a potentially nasty bug due to the pthread barrier
inside of the timers; if some threads exceed 0.5 and others do not, then
the program will hang indefinitely.

This will have to be resolved in the future by a mutex based on *all*
thread runtimes exceeding 0.5 seconds.  But that is currently beyond my
ability, so for now I just leave it as a stupid volatile mess.

Also, flops are now calculated locally, rather than outside of timer
completion.  This was necessary since the number of iterations is now an
unknown number.