Global runtime flag; barriers removed from timer
The thread-local runtime check was replaced with a global variable which
is set by individual threads.  This is probably not very thread-safe in
general, but my use of barriers may make this one safe.

Setting the flag is mutex-locked, which I suppose is the right thing to
do here, but I am not an experience async programmer.

The criterion is if any thread exceeds the 0.5s cutoff, then we
terminate the loop.

Timing info was also removed from the timer object, and moved explicitly
into the AVX code blocks.