TSC timer fix and re-enabled
Our TSC timer stop was (very) slightly off due to using `cpuid` to block
the OOO instruction execution rather than `rdtscp`, which blocks but
does not implicitly append the cpuid time.

We also move the `cpuid` instruction to occur after `rdtscp` to ensure
that no instructions will be inserted between the final loop
instructions and `rdtscp`.  (rdtscp ensures that all instructions will
complete prior to TSC measurement, but does not ensure that subsequent
instructions won't be preempted before measurement.)

Neither of these changes should have a measureable impact, since we are
removing a few cycles out of (potentially) millions or billions of
cycles, and we still cannot get a TSC frequency with a precision greater
than O(μsec).  But it is more correct with respec to serialization.