Roofline test input argument; r_max starts at 1
The roofline function is now an input argument, which will help us to
loop over all of the roofline tests.

This design is still very messy, but I'd rather get the iteration
working than worry about design at the moment.

We now start our loop iterator (r) at 1, rather than 1000.  The time to
reach ~1000 is negligible and this allows better support for much slower
loops, such as vectors in RAM.