r_max double race condition fix
In axpy, I was correctly using the local value of runtime to determine
if r_max should be doubled.  This was leading to cases where r_max would
be double if a different thread was below threshhold and would give
incorrect results.

Now, we explicitly check the mutex flag itself to determine if we should
double, which seems to have fixed the problem.