Redundant typedefs removed from stopwatch
I have since learned that it is bad practice to create a typedef to a
struct pointer for the sake of convenience, so I have removed these from
the stopwatch object.  I also removed typedefs around unions and enums.

I have kept the `Stopwatch` type since it's supposed to be a kind of
"class" for stopwatch "objects".

This might allow for additional simplication around the stopwatch
header.