"Score-P" mean TSC counter method
We now adopt the method from Score-P for measuring TSC frequency, which
takes the mean value of the TSC counter before and after time
measurements.  This is presumably more accurate, since the true value
will be somewhere inbetween the two values, and the mean is as good a
guess as any.

For some reason, our measurements appear to be about 10x worse if taken
mid-run than during the beginning.  (As in 10s of microseconds rather
than microseconds).  Currently I don't know how to prevent this, or what
could even be causing it.

But the current implementation seems to be an improvement, with TSC
values very close to the theoretical value of 2.6000... on Raijin.