TSC comments added, extended range test
Some explanation and attribution of the TSC timer asm code was added.

I also extended the clock_gettime cycle range check, since there was a
large discrepancy on my desktop.

To tell the truth, there is something more going on here, but I
ll have to leave it aside for now.