avx_mac cleanup
The avx_mac code is now iterable and does not do the old Alex Yee
add/subtract method.  We now just add and multiply by a small number
(1e-6, 1 + 1e-6) which is more robust to the weird AVX512 removal issue
(which is not correct, and may be an Intel bug, but nevermind that...)

I also found that explicit unroll pragmas needed to be applied for Intel
compilers, so those have now been added.