Framework for calculation bandwidth was added
We have added bandwidth output fields and reporting to the flop code.

The current numbers in avx_add are mostly nonsense, since data does not
really move on or off the register.  (It arguably moves through the CPU
ports, but I doubt that the bandwidth calculation is very meaningful.)

Sensible bandwidth values need to be added to the roofline tests.