bench_arg_t phaseout, single thread arg type
It was redundant and sort of useless to have a thread argument struct
which moved everything into a bench argument struct.  The whole
`bench.c` code was largely irrelevant.

This new version removes the generic thread input function and just uses
the benchmark functions themselves as thread inputs.  So far so good...

Bandwidth and flop output is now disabled if the value is zero (i.e. not
meaningful for the benchmark; might use a better method later.)