Roof speed with `restrict`
The `restrict` keyword was added to improve performance of the roofline
tests when using gcc.  Without the keyword, gcc was assuming the
possibility of pointer aliasing and doing additional work to preserve
the contents, which was slowing down the pipeline.

Intel icc appeared to assume no pointer aliasing and did not have this
issue.

Performance appears to have increased by about 10-15% in the
peak-bounded tests (axpy, axpby).