Bandwidth; roof test struct; mem alloc fix
We now compute bandwidths for the various roofline tests (and array
copy).  Numbers are a little different from stream (mostly higher) but
seem somewhat consistent.

(Arithmetic tests are a step higher than hardware, so I am not yet sure
how to interpret those.)

A new roof test struct for input arguments was created to accommodate
the bandwidth tests, which were added to the arguments.

Also added a free() for the roofline vectors; we were running out of
memory on the larger vector tests.