Better memory management
Stopwatches were not being deallocated by `free`, which was causing
crashes when iterating over multiple vector lengths (NYI).  This was
fixed by deallocating them in the Stopwatch `destroy` methods, but I am
starting to question this approach and may move the allocate/deallocate
outside of the function.  (Still not sure, because there is internal
detail to cleanup.)