Intel: remove copy unroll, re-enable memcpy
KNL was sort of disasterous without the avx512f_memcpy function, and I'm
not so concerned with getting well-vectorised copies, so I am restoring
memcpy for Intel builds.

It also turns out that explicit pragmas disable use of memcpy, so that
was also removed.  It helped, but did not produce "good" memcpy
performance on Sandy Bridge.