Replace CLOCK_MONOTONIC_RAW if not available
Older Linux kernels do not support CLOCK_MONOTONIC_RAW.  If this is not
available, then we replace it with the older CLOCK_MONOTONIC clock.

AFAIK, the only difference is that CLOCK_MONOTONIC is sensitive to NTP
time adjustments, which the RAW clock ignores.

A better check should be used here someday, such as introducing a new
clock variable and doing a kernel version check?  (Or using autoconf...)
Warning the user might also be nice.

In any case, this is a simple temporary solution.