Explicitly include `features.h` to get __USE_GNU
On Fujitsu SPARC, setting `_GNU_SOURCE` was not setting `__USE_CPU` for
some unknown reason.  My understanding is that all of the libc includes
are supposed to include `features.h`, but maybe some are missing.

Explicitly including `features.h` seemed to work on our SPARC platform,
and should not hurt on standard x86 Linux platforms.