Kernel wrapping prototype idea
This was an attempt to define kernel functions which update a single
element of a vector, and then apply those as inputs to a more general
wrapper function.  The idea was to eliminate all the repetition in
roof_copy, roof_axpy, etc.

Applying the kernel as an argument was hopeless.  I had much more
success defining explicit wrappers at compile-time, where the kernel was
provided as a constant argument to the wrapper inside multiple
functions.

While this mostly worked, I was still seeing slightly smaller
performance (about 5% lower), possibly because it was trying to do some
offset calculation at the zero-element.

This is too bad, but I have hope that there's still a good way to do
this, so I am holding on to this patch.