Macro-based roofline tests
The individual roofline tests are now generically produced by include of
the `roof.inc` template function.

I am not particularly proud of this, but I think it's better than copy
and pasting a function five (or more) times.

I have experimented with a few other methods (see kernel_wrap branch)
but nothing approaches peak performance as well as the explicit
implementation.

If there's a better way, even if it involves migration to C++, then I am
open to it.  But for now I will stick with this macro method.