Development code cleanup
Most of the early dev programs (axpy, kflop, xyzp) were moved to a
separate `dev` directory.  These mostly represent alternative methods
that I don't want to lose, or have not yet been incorporated into the
main program.

`axpy` is nearly identical to `flop`, but gives anomalously high
performance for 3200 when the `restrict` keyword is *not* used, and
anomalously low performance otherwise.  The assembly is nearly identical
across versions, so I am still not sure what the underlying cause could
be.

`kflop` was a weird attempt to inject a profiler into the kernel.

`xyzp` are triad operations not yet integrated into the main program.

`freq` was removed since it has been incorporated into the program.