Explicit cpu id list
We now explicitly determine the list of available CPUs.  Previously we
were assuming that the CPUs were contiguous and always avaiable.  We now
check each CPU to see if it's available, and append to a list if so.

This was required on one of our PowerPC platforms, which had various
CPUs turned off for some reason.