integer bugfix; overflow bugfix
Our loop-local nbench was a bust and had a bug, so reverting back to
function scope.

Also, longer runs were causing overflow in the flop and bandwidth roof
counters, so r (and r_max) are now saved as long.

There's probably a better solution here, but it will do for now.