Directory restructure; explicit vec test
The directory tree was renamed and moved around to accommodate the new
autoconf-generated Makefile.  Files are now stored in `src`, headers in
`inc`.

The vectorisation instruction tests are now done more explicitly,
although they still rely on standard libc headers (immintrin.h).

None of this is perfect, and it may be time to graduate to automake, but
it is much better than the hand-edited Makefile from before.