autoconf: tsc test bugfix; lib tests
We now use autoconf to search for external library dependencies,
specifically the `ceil` math function and the various `clock_*` timer
functions.  Often the latter does not require explicit linking to librt.

We also fixed a bug in the logic test for the assembly instruction flags
for the TSC timer method.