autoconf: Better Intel builds, streaming flags
Fixed an Intel build issue where optimisers were removing the test code.
This was was resolved by declaring the const register as volatile.  (How
is a const also volatile??)  This test may need to be revised or changed
in the future.

CFLAGS is now applied to every file, instead of just the vectorised
loops.  The hardcoded -O2 and -I. flags have been removed.

Added messages for enabling built-in functions and streaming stores.