autoconf: No need to check -std=gnu99 in $CC
Moving around platforms has shown that appending -std=gnu99 is not
predictable, and can no longer be included in the $CC tests.

We now use `read` to split off the compiler name from the arguments, and
exclusively test for gcc or icc in our various tests.

This resolved some autoconf issues on our KNL nodes.