Roofline kernels declared as static
The roofline kernels need to be local to the file in order to be
inlined, so should never be called outside of their files.  But since
they were not declared as static, compilers appeared to be making extra
standalone versions in the object file (just in case?).

By declaring them static, this should reduce the object file size by a
bit, and prevent calling of the functions outside of the roof.c file.