More stable tsc frequency calculation
The TSC timer is not very useful without an accurate TSC frequency
calculation.  Without access to the kernel, we must estimate it using
kernel clock functions (e.g. POSIX clock_gettime).

The TSC frequency calculation replaces the clock_nanosleep call with
aggressive loops of repeated clock_gettime calls.  This keeps
clock_gettime in CPU cache and ensures that execution time remains fast.

We lead with a 100k repeat cycle, which reduces the clock_gettime call
to about 200 cycles on my laptop (Kaby Lake?).  The actual time
measurement now does a fixed 10M repeat cycles, which does inert
calculations for about 1 second (0.7 on my laptop).

TODO: This iteration number will vary across platforms and needs to be
dynamically calculated (or estimated).

These changes appear to consistently get the frequency to about 1 part
on 10^7, or 100 ns, which is at least an order of magnitude of
improvement (maybe even 2, not sure).  This may be even higher, need to
do more tests.

The threshold test was reduced from 10x to 2x, although so far it is no
longer even necessary.

There may be a slight bias here, since the estimated frequency is always
below 2.4 GHz (again, on my laptop) but for all I know that could be
correct...