Makefile: Separate object file directory
The Makefile template was modified to store object files in a separate
directory, creating a slightly cleaner top directory.

GPU-related object files are still handled separately, but they can be
incorporated after the cuda path is determined by autoconf.

Some of the rules can also probably be simplified, but this is arguably
a separate problem.

----

Also, a minor note on the microbench documentation, noting that it was
written for an older version.