TSC freq determined before thread creation
Previously, when using a TSC-based timer, each test and thread would
independently determine its TSC frequency.  Since this can take a
significant amount of time, it was adding several seconds to the
runtime.

We now determine the TSC frequency prior to setting up the test (and
threads) and then explicitly set this frequency after Stopwatch creation
within the thread.

The implementation is not very elegant since we pass and set TSC
frequencies for both TSC and non-TSC timers (e.g. POSIX), so this still
needs to be cleaned up.  I'm still unsure how to best implement this.

NOTE: In the future we may need each core to independently determine its
own TSC frequency.