Stopwatch header cleanup (partial)
I was able to move the POSIX context definition out of the header and
into the .c file, but the TSC context is still accessed directly by some
tests and could not be removed.

I also removed a few components which are no longer used (tc_untyped and
the context_size array).