Global TSC frequency; Stopwatch interface cleanup
TSC frequency is now saved to a global variable, instead of being pass
around across threads into local timers.

"Globals are bad", but this is truly a global constant (in normal cases)
and I think it makes sense to be saved to a global variable.

This probably needs to be generalised at some point, maybe per-CPU, e.g.
if we ever extend this problem to multiple nodes, but for now I think
it's OK to keep it as a global.

This change allowed continued cleanup of the Stopwatch interface.  TSC
context is now defined locally in the source file.

A `stopwatch_get_tsc_freq` function was created to access the tsc
frequency, but maybe this is being a little pedantic.