TSC frequency removed from header
TSC frequency can be accessed indirectly, so it is not necessary to
expose the actual variable.  Also added a missing dummy function for the
generic case.

Note: these functions are still bad!  get_tsc_freq() should get the
variable, not call the calibration function.  Still need to clean all
this up.