Stopwatch interface cleanup; "rev up" block
The functions for separate stopwatch types were moved out of the
interface header and into separate clock-specific headers.  This further
protects internal details from users of Stopwatch.

I have also added a loop which arbitrarily spins for a fixed time
(currently 0.1 seconds) in order to rev up the frequency in frequency
scaling systems (i.e. Intel).  This provides more consistent results
between POSIX and TSC timers (which rev up by default) and also remove
any dud inital runs.

The loop is a bit blah and needs to be cleaned up.  At this point it's
just a proof of concept.  And the 0.1s cutoff is arbitrary and could use
some investigation.