Intel-specific pragma under conditional block
The explict `#pragma unroll` statements added to support Intel
performance were raising warning for some versions of gcc, and were
never intended to be parsed by gcc anyway, so we now keep these in a
conditional block.

Note that this sort of violates the principle of no preprocess
conditionals in non-header code.