Autoconf: Generalized asm,register,volatile
This is a very small start to the cleanup of the autoconf build, which
is currently far too hard-wired for general use.

In its current iteration, I tried to make the code build under GNU, and
then use autoconf to apply various modifiers to the build, particularly
under Intel.

In retrospect, this is a bit against the spirit of autoconf, which only
intervenes when it is necessary, rather than just convenient.

This actual PR does not do very much, other than introduce some macros
for the `asm` and `restrict` keywords, as well as a test for `volatile`,
and the latter two are arguably not even necessary.

Rather, this is just a small step towards cleaning up and organizing the
autoconf build.