C standard timer support; config.h integration
C99 builds don't support clock_gettime, so this patch adds the simpler
clock() timer as defined in the C standard library.

This new timer is now defined as the fallback timer, and is defined
inside of stopwatch.c, and the POSIX timer is now moved to an external
directory.

There is no compile-time configurations or selection of the timers, and
both are always built, so this is just an iterative step which includes
the timers but does not actually use them.

Good news is that the std C timer seems accurate enough to give the
right result for at least the first few digits.

---

Also did some small integration of config.h in order to pass the
definiton of `asm` for inline assembly.

And added a description to the AC_DEFINE to get autoheader working.