Tasks: Minimal implementation
This is a first attempt to integrate a generic Task object for managing
threads, or "tasks", or whatever.

These classes currently do nothing, and are only placeholders at the
moment.  The idea is that main() will create an array of Tasks, and then
use the API to join them to create things like mutexes and barriers, and
then finally fire them off as threads.

As generic as pthreads are, they are often unavailable on more esoteric
platforms, and we were starting to push weird flags and preprocessor
arguments in order to enable them.  Also, our thread management was
getting very messy, and hopefully this will start to organize them.

Placeholders for serial and POSIX (pthread) Tasks were created, but
neither of these do anything at the moment.