Restructure of roofline content in bench.h
The errant file bench.h is further whittled down by moving its
roofline-specific content to roof.h.

The last bit of content to be moved:

    - thread_args: Used to convert (void *) to (struct thread_args *)
      when passing arguments to pthread_create()

    - bench_roof_t: Used to define a list of tests

The second is related to the awkward merging of SIMD and roofline tests,
and is unlikely to survive its restructuring.

The first is harder, since it is setup in main.c and used in every test,
both roofline and SIMD, to convert the void pointer.

-

Both of these probably may be resovled by passing the roofline test as a
first-order pointer to pthread_create, rather than just passing it as an
argument to the SIMD tests.