Introduction of task struct
This is the first step to more cleanly organizing the runs which
separates the single list into a list of SIMD and Roofline tests, each
as a list of "task" structs.

This is very superficial, since we are still passing `roof_thread` and
doing the redundant pass-through from thread to roof_thread.  But this
lays the groundwork to directly pass.

This also makes a slightly fortuitous cleanup of the (admittedly very
bad) IO, which has these arbitrary indexing of `results[][]`.  This at
least starts on the roof tests, and doesn't need to keep getting updated
for new SIMD tests.  (Current values were even wrong!)