Header cleanup
This patch mostly just updates many of the header #define tags, which
contained many old labels from earlier versions of the program.

Also did a bit of a reshuffle of roof_thread and simd_thread to make
them appear equivalent (and possibly prepare for a merge into a single
function).