Roofline metrics moved to roof_args
The roofline metrics used to compute FLOP rates and bandwidths are
currently passed as hard-coded function arguments, which is a bit daft.

This patch moves these parameters into the `args` struct, and are now
specified inside of the kernel wrapper function.

SIMD tests could potentially use these someday if they were restructured
as kernels (but that seems unlikely due to complex initialization).