Temporary autoconf config fix
This temporarily stores the $CC variables in $ac_cc, to support some legacy
macros.

Not exactly sure what's happening here, perhaps autoconf used to set $ac_cc,
but dropped it in later versions.  In any case, the newest autoconf certainly
does *not* set $ac_cc, so there's no reason we should be using it.

Hopefully this will be fixed as part of an overall autoconf cleanup in the near
future.