Autoconf: Timer detection; optional POSIX timer
The main part of this patch is the autoconf timer function
configuration.  We now check for clockid_t and clock_gettime before
enabling the POSIX timer.  If unavailable, then we fallback to the C
standard timer.

The TSC timer is now only tested if the POSIX timer is also available.
This is because we used the POSIX timer to calibrate the TSC timer.  But
we might be able to use the C standard timer.

This resolves the C99 issues related to timers, but we still cannot
build in a pure C99 environment due to use of POSIX threads.