Compiler-agnostic vectorization tests
This patch removes the compiler-specific vectorization flags and
replaces them with agnostic tests for the vectorization intrisics
provided by Intel's `immintrin.h` header.

This could be problematic for the following reasons:

* A platform and compiler could support vector operations, even if the
  compiler does not support immintrin.h

* The user must now set explicitly set CFLAGS to enable vectorization if
  the compiler does not.

This shifts much of the burden onto the user, but overall I think it is
the better approach:

* We explicitly invoke the immintrin.h functions in our simd tests, so
  it makes more sense to explicitly test them.

* It is never a good idea to assume a compiler type, and to instead
  develop agnostic tests which confirm whether a particular feature
  (e.g. some SIMD type) is supported.

  Our old approach was even worse, since we were explicitly testing the
  name of CC.

Moving forward, it is possible to extend the tests to cycle through some
set of flags to confirm that some SIMD test is enabled.  For now, we
just check if the user turned it on themselves via CFLAGS.