Use nvcc to link the gpu build
The gpu build is a bit problematic, and needs to be integrated better to
autoconf.  But there is a `gpuflop` build which sort of works and sort
of used to work.

We used to use g++ to link the CUDA object code to the regular C code,
but that appears to be a bit problematic these days (something about
finding the cudart library).

This patch just uses nvcc for linking, and it seems to fix up these
issues.