SIMDTYPE type macro and conversion
This introduces the SIMDTYPE preprocessor macro, which defaults to
`float` but can be overridden to `double`.

Maybe someday I can have both, but this is an easy enough first step.
The need for both float and double is just getting too messy to ignore
anymore.

Also removed `bench.h` from the SIMD headers since only `roof.h` was
needed.

Finally, this generalizes much of the float-specific content in those
SIMD functions.