GPU SIMD generalization
Adding the SIMDTYPE macro the gpu simd test

In order to pass DEFS to both C and CUDA, the CPPFLAGS macro is now
used.  It is filled with DEFS (which just points to config.h) as well as
user-defined CPP defs (i.e. SIMDTYPE).

The include directory was also moved to this variable, since it's used
by both C and CUDA.