Remove intrinsincs for AVX tests
Seems like the intrinsic calls are not needed, and the compiler
correctly flags them as vectorizable and localized to registers.

Updated tests:
- avx_add
- avx_mul
- avx_mac

Note that avx_mac performance has dropped on my Ryzen 5.  It is possible
that I was never actually doing it properly.

The issue seems to be due to a lack of registers.  This might not have
been an issue with the intrinsics.  And maybe also means this is not
actually what I want to be doing.