Improved frequency measurement for latency test
Latency testing requires knowledge of the CPU frequency.  There were
many problems with the existing test:

1. The frequency changes over time, the existing test was not pushing
   the CPU into its "turbo"/"boost" state on most machines.

2. The `v == 1` instruction of the loop was not really doing anything,
   and seems to even be (understandably) removed in most optimized
   builds.  To make things worse, the (empty) loop appeared to be
   executing at double time on some procs, reporting a doubled
   frequency.

   This is replaced with an update to a volatile character variable.
   It appears to be a 1-cycle operation which throttles the loop (whose
   other instructions hopefully execute in parallel) and appears to
   produce a valid frequency on a wider range of platforms.

The subsequent loops used to compute latency are now in a turbo state
and do not need to run as long, so the `r_max` is significantly reduced
for these tests.

Overall, this new version appears to be giving much more accurate
estimates for latency, and this may soon be ready for integration into
the model.