Remove per-thread output for single thread
Reporting per-thread output for single-thread jobs was not particularly
helpful, and only added to the already messy model output.  This is now
only reported if there is more than one thread.

Bandwidth output is now removed for SIMD-only output.  Although a
theoretical bandwidth can be assigned to the register-L1 cache, it's not
supposed to be a factor in these calculations.