Partial step towards including reference BLAS test
This adds the reference Fortran implementation of BLAS's DGEMM function
to the repository, as well as rules and configuration macros for
building the source.

It is not possible to run the function yet, and if one were to run it
then it would probably give very incorrect answers.  This should only be
considered a partial commit.  I expect to later replace it with a
working version.