Only use AC_FC_LIBRARY_LDFLAGS on newer autoconf
The AC_FC_LIBRARY_LDFLAGS (used for mixed C-Fortran compilation) only
really works if libtool is also set up, which is very much not the case.

Newer autoconfs can safely just disable it if the necessary libtool
content is missing, but older autoconfs do not handle it very well and
cause errors.

Although not an ideal solution, this patch disables the macro for older
versions of autoconf.  We will come back to this problem later, when the
problem is better understood.