Remove the autoconf loop unroll configuration
This removes the loop unrolling flag configuration from autoconf.  Aside
from using flags which are only specific to GNU (and apparently
tolerated by Intel), it is yet another example of configuration which
should probably be handled by the user rather than autoconf.

Maybe someday we can identify robust tests for this sort of thing, but
its current form was seriously flawed and needed to go.

Delimiters were also added to the libm test.