Nvidia/CUDA support overhaul
Several changes were made to enable smoother support for CUDA-based
testing.

The most significant changes were related to autoconf and Makefile
generation.  Linking by nvc++ was dropped and we once again link with
the CUDA compiler.  For unclear reasons, linking with C++ seemed to have
an adverse effect on build times.  It also avoided the need to locate
the CUDA library paths.

We cannot use the CUDA compiler to build everything, as in some previous
attempts.  This particularly breaks down for the SIMD testing, with
errors related to the FMA intrinsics.  So in general we still use nvc to
build object code.  The exception is main.o, which seems to otherwise
use functions unusable by the CUDA compiler.  To get around this, we
compile main.c with the CUDA compiler.  (It doesn't make much sense me,
but it works...).  This is handled with the MAINCC macro.

Some other minor tweaks:
* PTHREAD is now a macro used to control pthread flags, partially
  configured by autoconf.  Notably, Nvidia does not need nor does it
  recognize a -pthread flag.

* the cuBLAS library configuration is now handled by autoconf, rather
  than dumdum LIBS manipulation.

* The old platform-specific Fujitsu compiler stuff was finally purged
  from the old Makefile.in template.  It's not how we expect things to
  work anymore.  (Old comments written by "someone" who did not
  understand the nuances of the -march flag have also been removed.)

* TSC and CBLAS path configuration has been simplified.

* The default min_runtime was increased from 0.01 to 0.1, since the
  latter value was far more favorable to GPUs.

  Maybe this was too aggressive, but better to be safe than sorry.  We
  can talk about target-specific defaults when this project is important
  enough to be burdened by such things.