Redefine Stopwatch funcs with prototypes
The function pointers in Stopwatch had been defined with incorrect
prototypes.  This is because it would have required a reference to
Stopwatch inside of the Stopwatch struct definition, which is not
possible.  So I just left them empty (and incorrect).

This is apparently no longer permitted in C2x, and modern compilers were
raising objections.

The way out of this paradox is to define the type before defining the
struct.  Who knew that was an option?  (Better programmers than me,
obviously!)

We now separate the typedef from the struct, and lead with the typedef.
This allows us to reference the typedef when defining the struct.
(self-reference?)